PM Evaluation updates & vendor account page

+++ application/views/admin_new/languages.php
@@ -1,19 +1,3 @@
-<?php if ($this->session->flashdata('true')) { ?>
-  <div class="alert alert-success" role="alert">
-    <span class="fa fa-check-circle"></span>
-    <span><strong>
-        <?= $this->session->flashdata('true') ?>
-      </strong></span>
-  </div>
-<?php } ?>
-<?php if ($this->session->flashdata('error')) { ?>
-  <div class="alert alert-danger" role="alert">
-    <span class="fa fa-warning"></span>
-    <span><strong>
-        <?= $this->session->flashdata('error') ?>
-      </strong></span>
-  </div>
-<?php } ?>
 <!--begin::Content-->
 <div class="content d-flex flex-column flex-column-fluid" id="kt_content">
 
@@ -21,43 +5,50 @@
   <div class="d-flex flex-column-fluid">
     <!--begin::Container-->
     <div class="container">
-      <?php if (isset($_SERVER['HTTP_REFERER'])): ?>
-        <input type="text" name="referer" value="<?= $_SERVER['HTTP_REFERER'] ?>" hidden>
-      <?php else: ?>
-        <input type="text" name="referer" value="<?= base_url() ?>account" hidden>
-      <?php endif; ?>
+
+      <?php if ($this->session->flashdata('true')) { ?>
+        <div class="alert alert-success" role="alert">
+          <span class="fa fa-check-circle"></span>
+          <span><strong><?= $this->session->flashdata('true') ?></strong></span>
+        </div>
+      <?php  } ?>
+      <?php if ($this->session->flashdata('error')) { ?>
+        <div class="alert alert-danger" role="alert">
+          <span class="fa fa-warning"></span>
+          <span><strong><?= $this->session->flashdata('error') ?></strong></span>
+        </div>
+      <?php  } ?>
       <!-- start search form card -->
       <div class="card card-custom gutter-b example example-compact">
         <div class="card-header">
           <h3 class="card-title">Search Languages</h3>
         </div>
 
-        <form class="form" id="languages" action="<?php echo base_url() ?>admin/languages" method="post"
-          enctype="multipart/form-data">
+        <form class="form" id="languages" action="<?php echo base_url() ?>admin/languages" method="get" enctype="multipart/form-data">
           <div class="card-body">
+
             <div class="form-group row">
 
               <label class="col-lg-2 col-form-label text-lg-right">Language</label>
               <div class="col-lg-3">
-                <input type="text" class="form-control" name="lang_name" id="lang_name"
-                  value="<?= $_SESSION['lang_name'] ?? '' ?>">
+                <input type="text" class="form-control" name="name" value="<?= $name ?? '' ?>">
+
               </div>
+
             </div>
 
             <div class="form-group row">
 
               <label class="col-lg-2 col-form-label text-lg-right">Date From</label>
               <div class="col-lg-3">
-                <input class="form-control date_sheet" type="text" name="date_from" autocomplete="off"
-                  value="<?= $date_from ?? '' ?>">
+                <input class="form-control date_sheet" type="text" name="date_from" autocomplete="off" value="<?= $date_from ?? '' ?>">
 
               </div>
 
-              <label class="col-lg-2 control-label" for="role date_to">Date To</label>
+              <label class="col-lg-2 control-label" for="role name">Date To</label>
 
               <div class="col-lg-3">
-                <input class="form-control date_sheet" type="text" name="date_to" autocomplete="off"
-                  value="<?= $date_to ?? '' ?>">
+                <input class="form-control date_sheet" type="text" name="date_to" autocomplete="off" value="<?= $date_to ?? '' ?>">
 
               </div>
 
@@ -68,16 +59,9 @@
               <div class="row">
                 <div class="col-lg-2"></div>
                 <div class="col-lg-10">
-                  <button class="btn btn-success mr-2" name="search" type="submit" value="search"><i
-                      class="la la-search"></i>Search</button>
-                  <button class="btn btn-warning" type="submit" name="submitReset" value="submitReset"><i
-                      class="la la-trash"></i>Clear
-                    Filter</button>
-                  <button class="btn btn-secondary"
-                    onclick="var e2 = document.getElementById('languages'); e2.action='<?= base_url() ?>admin/exportLanguages'; e2.submit();"
-                    name="export" type="submit"><i class="la la-download" aria-hidden="true"></i> Export To
-                    Excel</button>
-
+                  <button class="btn btn-success mr-2" name="search" onclick="var e2 = document.getElementById('languages'); e2.action='<?= base_url() ?>admin/languages'; e2.submit();" type="submit">Search</button>
+                  <button class="btn btn-secondary" onclick="var e2 = document.getElementById('languages'); e2.action='<?= base_url() ?>admin/exportLanguages'; e2.submit();" name="export" type="submit"><i class="fa fa-download" aria-hidden="true"></i> Export To Excel</button>
+                  <a href="<?= base_url() ?>admin/languages" class="btn btn-warning"><i class="la la-trash"></i>Clear Filter</a>
 
                 </div>
               </div>
@@ -101,19 +85,16 @@
             <a href="<?= base_url() ?>admin/addLanguage" class="btn btn-primary font-weight-bolder">
             <?php } ?>
             <span class="svg-icon svg-icon-md">
-              <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px"
-                height="24px" viewBox="0 0 24 24" version="1.1">
+              <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px" height="24px" viewBox="0 0 24 24" version="1.1">
                 <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
                   <rect x="0" y="0" width="24" height="24" />
                   <circle fill="#000000" cx="9" cy="15" r="6" />
-                  <path
-                    d="M8.8012943,7.00241953 C9.83837775,5.20768121 11.7781543,4 14,4 C17.3137085,4 20,6.6862915 20,10 C20,12.2218457 18.7923188,14.1616223 16.9975805,15.1987057 C16.9991904,15.1326658 17,15.0664274 17,15 C17,10.581722 13.418278,7 9,7 C8.93357256,7 8.86733422,7.00080962 8.8012943,7.00241953 Z"
-                    fill="#000000" opacity="0.3" />
+                  <path d="M8.8012943,7.00241953 C9.83837775,5.20768121 11.7781543,4 14,4 C17.3137085,4 20,6.6862915 20,10 C20,12.2218457 18.7923188,14.1616223 16.9975805,15.1987057 C16.9991904,15.1326658 17,15.0664274 17,15 C17,10.581722 13.418278,7 9,7 C8.93357256,7 8.86733422,7.00080962 8.8012943,7.00241953 Z" fill="#000000" opacity="0.3" />
                 </g>
               </svg>
               <!--end::Svg Icon-->
             </span>Add New Language</a>
-          <!--end::Button-->
+            <!--end::Button-->
         </div>
       </div>
       <div class="card-body">
@@ -134,20 +115,12 @@
             <?php
             if ($languages->num_rows() > 0) {
               foreach ($languages->result() as $row) {
-                ?>
+            ?>
                 <tr class="">
-                  <td>
-                    <?php echo $row->id; ?>
-                  </td>
-                  <td>
-                    <?php echo $row->name; ?>
-                  </td>
-                  <td>
-                    <?php echo $this->admin_model->getAdmin($row->created_by); ?>
-                  </td>
-                  <td>
-                    <?= $row->created_at ?>
-                  </td>
+                  <td><?php echo $row->id; ?></td>
+                  <td><?php echo $row->name; ?></td>
+                  <td><?php echo $this->admin_model->getAdmin($row->created_by); ?></td>
+                  <td><?= $row->created_at ?></td>
                   <td>
                     <?php if ($permission->edit == 1) { ?>
                       <a href="<?php echo base_url() ?>admin/editLanguage?t=<?php echo base64_encode($row->id); ?>" class="">
@@ -158,23 +131,20 @@
 
                   <td>
                     <?php if ($permission->delete == 1) { ?>
-                      <a href="<?php echo base_url() ?>admin/deleteLanguage/<?php echo $row->id ?>" title="delete" class=""
-                        onclick="return confirm('Are you sure you want to delete this language?');">
+                      <a href="<?php echo base_url() ?>admin/deleteLanguage/<?php echo $row->id ?>" title="delete" class="" onclick="return confirm('Are you sure you want to delete this language?');">
                         <i class="fa fa-times text-danger text"></i> Delete
                       </a>
                     <?php } ?>
                   </td>
                 </tr>
-                <?php
+              <?php
               }
             } else {
-              ?>
-            <tr>
-              <td colspan="7">There is no Languages to list</td>
-            </tr>
-            <?php
-            }
-            ?>
+              ?><tr>
+                <td colspan="7">There is no Languages to list</td>
+              </tr><?php
+                  }
+                    ?>
           </tbody>
 
         </table>
@@ -190,5 +160,6 @@
   </div>
   <!--end::Container-->
 </div>
-
+<!--end::Entry-->
+</div>
 <!--end::Content-->
